flask-code/requirements.txt
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cohere==3.1.3
dnspython==2.2.1
Flask==2.2.2
idna==3.4
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.6.0
MarkupSafe==2.1.1
pinecone-client==2.0.13
pycodestyle==2.10.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
six==1.16.0
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
